move metadata operate to InstanceController.
